refactor(sidekick): move files to internal/sidekick/sidekick
As a step toward consolidating the internal/sidekick/internal/ packages
under internal/sidekick, move the existing files into
internal/sidekick/sidekick.

Pure code in motion.
