+++ internal/sidekick/sidekick/cmdline.go
