+++ internal/sidekick/sidekick/cmdline_test.go
