+++ internal/sidekick/sidekick/command.go
