+++ internal/sidekick/sidekick/command_test.go
