+++ internal/sidekick/sidekick/downloads_cache.go
