+++ internal/sidekick/sidekick/downloads_cache_test.go
