+++ internal/sidekick/sidekick/generate.go
