+++ internal/sidekick/sidekick/refresh.go
