+++ internal/sidekick/sidekick/refreshall.go
