+++ internal/sidekick/sidekick/rust_generate.go
