+++ internal/sidekick/sidekick/sidekick_rust_prost_test.go
