+++ internal/sidekick/sidekick/update.go
